Fix: resolve 5 review bugs in UI redesign
- EntityForm: readOnly fields now visible in Others tab (section filter uses full fields array)
- DetailView Badge: merge className to preserve emerald status colors
- DetailView useEffect: add missing dependency array values
- Search bars: mark as readOnly placeholder until wired up
- Extract getStatusBadgeProps to shared lib/statusBadge.js (was duplicated with diverging logic)
